+++ honeybee_vtk/cli/__init__.py
@@ -4,7 +4,6 @@ import click
 from .translate import translate
 from .export import export
 from .config import config
-from .time_step_data import time_step_data
 from .post_process import post_process
 from honeybee.cli import main
 
@@ -21,7 +20,6 @@ def vtk():
 vtk.add_command(translate)
 vtk.add_command(export)
 vtk.add_command(config)
-vtk.add_command(time_step_data)
 vtk.add_command(post_process)
 
 main.add_command(vtk)
