+++ honeybee_vtk/cli/post_process.py
@@ -8,7 +8,7 @@ from honeybee_vtk.image_processing import write_gif, write_transparent_images
 
 @click.group()
 def post_process():
-    """Honeybee VTK command entry point."""
+    """Command to post-process media generated by honeybee-vtk."""
     pass
